moved duplicated code into a helpers script
